Updates https://github.com/pcca-matrix/PCCA-Layout/issues/68
Set the default wheel position to the center of the surface.
This prevents elements from going out of bounds when a large center zoom is set
add alignment marks when on edit mode to view where the center of the surface is.
fix special artworks not displayed (alpha not set)
animate pointer only when wheel is stopped.