Adding kiosk mode
Following an idea from triplesix974, adding a kiosk mode.
https://github.com/pcca-matrix/PCCA-Layout/issues/69
The kiosk mode prevents your setup from being altered when your system is in the hands of children or guests.
To activate kiosk mode, simply place a file named "kiosk" without an extension in the root of the theme next to the layout.nut file.
The options for kiosk mode are found in the pcca menu -> kiosk, and the options are:
Exit AM: blocks the ability to exit AM
Tags: blocks the ability to add/remove tags
Favorites: blocks the ability to add/remove favorites

If kiosk mode is activated and the option to add tags is set to Yes, you will no longer be able to create new tags but only add a game to an existing tag that is not present in a filter.

Useful for tags like ("not working", "completed", etc...) that you could leave the choice to tag the game with but prevent the modification of any lists of games created with others tags.